requirements.txt
click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
flask-jsontools==0.1.7
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2==2.9.1
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.27
Werkzeug==2.0.2
